seprates out query runner
