+++ src/py/crankshaft/crankshaft/clustering/moran.py
@@ -8,20 +8,12 @@ Moran's I geostatistics (global clustering & outliers presence)
 import pysal as ps
 import plpy
 from collections import OrderedDict
+from crankshaft.query_runner import QueryRunner
 
 # crankshaft module
 import crankshaft.pysal_utils as pu
 
 # High level interface ---------------------------------------
-
-
-class QueryRunner:
-    def get_result(self, query):
-        try:
-            data = plpy.execute(query)
-        except plpy.SPIError, err:
-            plpy.error("k-means (spatial) cluster analysis failed: %s" % err)
-        return data
 
 
 class Moran:
@@ -47,7 +39,7 @@ class Moran:
 
         query = pu.construct_neighbor_query(w_type, qvals)
 
-        result = self.query_runner.get_result(query)
+        result = self.query_runner.get_moran(query)
 
         # collect attributes
         attr_vals = pu.get_attributes(result)
@@ -79,7 +71,7 @@ class Moran:
 
         query = pu.construct_neighbor_query(w_type, qvals)
 
-        result = self.query_runner.get_result(query)
+        result = self.query_runner.get_moran(query)
 
         attr_vals = pu.get_attributes(result)
         weight = pu.get_weight(result, w_type, num_ngbrs)
@@ -108,7 +100,7 @@ class Moran:
 
         query = pu.construct_neighbor_query(w_type, qvals)
 
-        result = self.query_runner.get_result(query)
+        result = self.query_runner.get_moran(query)
 
         # collect attributes
         numer = pu.get_attributes(result, 1)
@@ -140,7 +132,7 @@ class Moran:
 
         query = pu.construct_neighbor_query(w_type, qvals)
 
-        result = self.query_runner.get_result(query)
+        result = self.query_runner.get_moran(query)
 
         # collect attributes
         numer = pu.get_attributes(result, 1)
@@ -173,7 +165,7 @@ class Moran:
 
         query = pu.construct_neighbor_query(w_type, qvals)
 
-        result = self.query_runner.get_result(query)
+        result = self.query_runner.get_moran(query)
 
         # collect attributes
         attr1_vals = pu.get_attributes(result, 1)
